# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FormationOut": {
        "properties": {
          "basin": {
            "title": "Basin",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "lithology": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lithology"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "strat_order": {
            "title": "Strat Order",
            "type": "integer"
          },
          "synonyms": {
            "items": {
              "type": "string"
            },
            "title": "Synonyms",
            "type": "array"
          }
        },
        "required": [
          "id",
          "basin",
          "name",
          "synonyms",
          "strat_order",
          "lithology"
        ],
        "title": "FormationOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProximityMode": {
        "enum": [
          "SURFACE",
          "AT_FORMATION",
          "CLOSEST_APPROACH"
        ],
        "title": "ProximityMode",
        "type": "string"
      },
...
      "RiskMode": {
        "enum": [
          "AT_FORMATION",
          "SURFACE"
        ],
        "title": "RiskMode",
        "type": "string"
      },
...
      "WellList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WellSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "WellList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "eRTMAC-NWIS offset-well knowledge and decision support (SIH PS 121)",
    "title": "SMRITI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/formations": {
      "get": {
        "operationId": "get_formations_api_v1_formations_get",
        "parameters": [
          {
            "in": "query",
            "name": "basin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Basin"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FormationOut"
                  },
                  "title": "Response Get Formations Api V1 Formations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Formation dictionary",
        "tags": [
          "wells"
        ]
      }
    },
...
    "/api/v1/reports/offset-brief/{well_id}": {
      "get": {
        "operationId": "offset_brief_api_v1_reports_offset_brief__well_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Planned, not implemented in this phase"
          }
        },
        "summary": "Offset Risk Brief (PDF)",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/v1/wells": {
      "get": {
        "operationId": "get_wells_api_v1_wells_get",
        "parameters": [
          {
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Field"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "description": "Map viewport 'min_lon,min_lat,max_lon,max_lat' (WGS84 degrees)",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Map viewport 'min_lon,min_lat,max_lon,max_lat' (WGS84 degrees)",
              "title": "Bbox"
            }
          },
          {
            "in": "query",
            "name": "fluid_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "oil",
                    "gas",
                    "water"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fluid Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WellList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List and filter wells",
        "tags": [
          "wells"
        ]
      }
    },
    "/api/v1/wells/{well_id}": {
      "get": {
        "operationId": "get_well_api_v1_wells__well_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WellDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Well 360 summary",
        "tags": [
          "wells"
        ]
      }
    },
    "/api/v1/wells/{well_id}/cementing-check": {
      "get": {
        "description": "Planned slurry density against the mud weights at which offsets lost circulation in\nthe formation at the shoe, and offsets' losses while cementing there.",
        "operationId": "get_cementing_check_api_v1_wells__well_id__cementing_check_get",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "shoe_md_m",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "maximum": 15000,
              "title": "Shoe Md M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "slurry_density_sg",
            "required": true,
            "schema": {
              "maximum": 2.6,
              "minimum": 1.0,
              "title": "Slurry Density Sg",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "default": 10.0,
              "exclusiveMinimum": 0,
              "maximum": 50,
              "title": "Radius Km",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CementingCheck"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cementing checklist from offsets",
        "tags": [
          "risk"
        ]
      }
    },
...
    "/api/v1/wells/{well_id}/offsets": {
      "get": {
        "description": "``SURFACE``: wellhead-to-wellhead distance. ``AT_FORMATION`` (needs ``formation``):\ndistance between the two wells' entry points into that formation; wells that never\nreach it are listed in ``excluded``. ``CLOSEST_APPROACH``: minimum 3D distance between\nthe wellbore paths inside the optional TVDSS window. ``radius_km`` bounds the distance in\nevery mode.",
        "operationId": "get_offsets_api_v1_wells__well_id__offsets_get",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "default": 5.0,
              "exclusiveMinimum": 0,
              "maximum": 100,
              "title": "Radius Km",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ProximityMode",
              "default": "SURFACE"
            }
          },
          {
            "description": "AT_FORMATION: formation to compare at",
            "in": "query",
            "name": "formation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AT_FORMATION: formation to compare at",
              "title": "Formation"
            }
          },
          {
            "description": "CLOSEST_APPROACH: TVDSS window start (m)",
            "in": "query",
            "name": "tvdss_from_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "CLOSEST_APPROACH: TVDSS window start (m)",
              "title": "Tvdss From M"
            }
          },
          {
            "description": "CLOSEST_APPROACH: TVDSS window end (m)",
            "in": "query",
            "name": "tvdss_to_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "CLOSEST_APPROACH: TVDSS window end (m)",
              "title": "Tvdss To M"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OffsetsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Offset wells within a radius",
        "tags": [
          "wells"
        ]
      }
    },
...
    "/api/v1/wells/{well_id}/risk-profile": {
      "get": {
        "description": "For each formation of the well: P(event | formation) from offsets within\n``radius_km`` (weighted Beta-Binomial with a basin prior), its 90% credible interval,\nn_eff and the contributing offsets. The well's own events are never used.",
        "operationId": "get_risk_profile_api_v1_wells__well_id__risk_profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "default": 10.0,
              "exclusiveMinimum": 0,
              "maximum": 50,
              "title": "Radius Km",
              "type": "number"
            }
          },
          {
            "description": "Spatial scale; default radius_km / 2",
            "in": "query",
            "name": "sigma_km",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 50,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Spatial scale; default radius_km / 2",
              "title": "Sigma Km"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RiskMode",
              "default": "AT_FORMATION"
            }
          },
          {
            "description": "Repeat for several; default: all seen",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "LOSS",
                      "KICK",
                      "STUCK",
                      "TIGHT",
                      "TORQUE",
                      "INSTAB",
                      "BALLING",
                      "OVERP",
                      "GAS",
                      "CEMENT",
                      "CASING",
                      "FISH",
                      "EQUIP",
                      "WAIT",
                      "OTHER_NPT"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeat for several; default: all seen",
              "title": "Event Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Offset prior risk by formation",
        "tags": [
          "risk"
        ]
      }
    },
    "/api/v1/wells/{well_id}/trajectory": {
      "get": {
        "operationId": "get_trajectory_api_v1_wells__well_id__trajectory_get",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Survey stations and 3D path",
        "tags": [
          "wells"
        ]
      },
      "post": {
        "description": "Replaces the primary wellbore's stations, recomputes them by minimum curvature,\nrebuilds ``path_geom`` and formation entry points, and clears ``trajectory_assumed``.\nDepth references derived from the trajectory (event, casing, cement and mud TVD/TVDSS)\nare recomputed too.",
        "operationId": "upload_survey_api_v1_wells__well_id__trajectory_post",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload survey stations and recompute the trajectory",
        "tags": [
          "wells"
        ]
      }
    },
    "/api/v1/wells/{well_id}/trajectory/at-depth": {
      "get": {
        "description": "Minimum-curvature interpolation between the bracketing stations; 422 when ``md_m``\nis beyond the last station.",
        "operationId": "get_trajectory_at_depth_api_v1_wells__well_id__trajectory_at_depth_get",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "md_m",
            "required": true,
            "schema": {
              "maximum": 15000,
              "minimum": 0,
              "title": "Md M",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryAtDepth"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Interpolated position at a measured depth",
        "tags": [
          "wells"
        ]
      }
    },
...
  }
}